Rename file slice-zone.vue to SliceZone.vue
